refactor: split DynamicPropertiesPanel into focused sub-components — reduce 699-line catch-all
Extract PropertyValueCells, TypeSelector, and AddPropertyForm into dedicated
files. Move shared display-mode constants to utils/propertyTypes to satisfy
react-refresh lint rule. All tests pass, no behavior change.
