+++ src/components/AddPropertyForm.tsx
@@ -1,0 +1,175 @@
+import { useState } from 'react'
+import { Button } from '@/components/ui/button'
+import { Calendar } from '@/components/ui/calendar'
+import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
+import { CalendarIcon, Check, X } from 'lucide-react'
+import {
+  type PropertyDisplayMode,
+  formatDateValue,
+  toISODate,
+} from '../utils/propertyTypes'
+import { StatusPill, StatusDropdown } from './StatusDropdown'
+import { DISPLAY_MODE_OPTIONS, DISPLAY_MODE_ICONS } from '../utils/propertyTypes'
+
+function parseDateValue(value: string): Date | undefined {
+  const iso = toISODate(value)
+  const d = new Date(iso + 'T00:00:00')
+  return isNaN(d.getTime()) ? undefined : d
+}
+
+function dateToISO(day: Date): string {
+  const yyyy = day.getFullYear()
+  const mm = String(day.getMonth() + 1).padStart(2, '0')
+  const dd = String(day.getDate()).padStart(2, '0')
+  return `${yyyy}-${mm}-${dd}`
+}
+
+const ADD_INPUT_CLASS = "h-[26px] min-w-[60px] flex-1 rounded border border-border bg-muted px-1.5 text-[12px] text-foreground outline-none focus:border-primary"
+
+function AddBooleanInput({ value, onChange }: { value: string; onChange: (v: string) => void }) {
+  const boolVal = value.toLowerCase() === 'true'
+  return (
+    <button
+      className="h-[26px] min-w-[60px] flex-1 rounded border border-border bg-muted px-1.5 text-[12px] text-secondary-foreground transition-colors hover:bg-accent"
+      onClick={() => onChange(boolVal ? 'false' : 'true')}
+      data-testid="add-property-boolean-toggle"
+    >
+      {boolVal ? '\u2713 Yes' : '\u2717 No'}
+    </button>
+  )
+}
+
+function AddDateInput({ value, onChange }: { value: string; onChange: (v: string) => void }) {
+  const selectedDate = value ? parseDateValue(value) : undefined
+  const formatted = value ? formatDateValue(value) : ''
+  return (
+    <Popover>
+      <PopoverTrigger asChild>
+        <button
+          className="inline-flex h-[26px] min-w-[60px] flex-1 cursor-pointer items-center gap-1 rounded border border-border bg-muted px-1.5 text-[12px] transition-colors hover:bg-accent"
+          data-testid="add-property-date-trigger"
+        >
+          <CalendarIcon className="size-3 shrink-0 text-muted-foreground" />
+          <span className={`min-w-0 truncate${!formatted ? ' text-muted-foreground' : ' text-foreground'}`}>
+            {formatted || 'Pick a date\u2026'}
+          </span>
+        </button>
+      </PopoverTrigger>
+      <PopoverContent className="w-auto p-0" align="start" side="left">
+        <Calendar
+          mode="single"
+          selected={selectedDate}
+          onSelect={(day) => { if (day) onChange(dateToISO(day)) }}
+          defaultMonth={selectedDate}
+        />
+      </PopoverContent>
+    </Popover>
+  )
+}
+
+function AddStatusInput({ value, onChange, vaultStatuses }: { value: string; onChange: (v: string) => void; vaultStatuses: string[] }) {
+  const [showDropdown, setShowDropdown] = useState(false)
+  return (
+    <span className="relative inline-flex min-w-[60px] flex-1 items-center">
+      <button
+        className="inline-flex h-[26px] min-w-[60px] flex-1 cursor-pointer items-center gap-1 rounded border border-border bg-muted px-1.5 text-[12px] transition-colors hover:bg-accent"
+        onClick={() => setShowDropdown(true)}
+        data-testid="add-property-status-trigger"
+      >
+        {value ? <StatusPill status={value} /> : <span className="text-muted-foreground">Status{'\u2026'}</span>}
+      </button>
+      {showDropdown && (
+        <StatusDropdown
+          value={value}
+          vaultStatuses={vaultStatuses}
+          onSave={(v) => { onChange(v); setShowDropdown(false) }}
+          onCancel={() => setShowDropdown(false)}
+        />
+      )}
+    </span>
+  )
+}
+
+function AddPropertyValueInput({ displayMode, value, onChange, onKeyDown, vaultStatuses }: {
+  displayMode: PropertyDisplayMode; value: string; onChange: (v: string) => void
+  onKeyDown: (e: React.KeyboardEvent) => void; vaultStatuses: string[]
+}) {
+  switch (displayMode) {
+    case 'boolean': return <AddBooleanInput value={value} onChange={onChange} />
+    case 'date': return <AddDateInput value={value} onChange={onChange} />
+    case 'status': return <AddStatusInput value={value} onChange={onChange} vaultStatuses={vaultStatuses} />
+    case 'tags': return (
+      <input className={ADD_INPUT_CLASS} type="text" placeholder="tag1, tag2, ..." value={value}
+        onChange={(e) => onChange(e.target.value)} onKeyDown={onKeyDown}
+      />
+    )
+    default: return (
+      <input className={ADD_INPUT_CLASS} type="text" placeholder="Value" value={value}
+        onChange={(e) => onChange(e.target.value)} onKeyDown={onKeyDown}
+      />
+    )
+  }
+}
+
+export function AddPropertyForm({ onAdd, onCancel, vaultStatuses }: {
+  onAdd: (key: string, value: string, displayMode: PropertyDisplayMode) => void; onCancel: () => void
+  vaultStatuses: string[]
+}) {
+  const [newKey, setNewKey] = useState('')
+  const [newValue, setNewValue] = useState('')
+  const [displayMode, setDisplayMode] = useState<PropertyDisplayMode>('text')
+
+  const handleModeChange = (mode: PropertyDisplayMode) => {
+    setDisplayMode(mode)
+    if (mode === 'boolean') setNewValue('false')
+    else if (mode !== displayMode) setNewValue('')
+  }
+
+  const handleKeyDown = (e: React.KeyboardEvent) => {
+    if (e.key === 'Enter' && newKey.trim()) onAdd(newKey, newValue, displayMode)
+    else if (e.key === 'Escape') onCancel()
+  }
+
+  return (
+    <div className="mt-1 flex flex-wrap items-center gap-1.5 rounded px-1.5 py-1" data-testid="add-property-form">
+      <input
+        className="h-[26px] w-20 shrink-0 rounded border border-border bg-muted px-1.5 text-[12px] text-foreground outline-none focus:border-primary"
+        type="text" placeholder="Property name" value={newKey}
+        onChange={(e) => setNewKey(e.target.value)} onKeyDown={handleKeyDown} autoFocus
+      />
+      <Select value={displayMode} onValueChange={(v) => handleModeChange(v as PropertyDisplayMode)}>
+        <SelectTrigger
+          size="sm"
+          className="h-[26px] w-[72px] shrink-0 gap-1 border-border bg-muted px-1.5 py-0 shadow-none"
+          style={{ fontSize: 12, borderRadius: 4 }}
+          data-testid="add-property-type-trigger"
+        >
+          <SelectValue />
+        </SelectTrigger>
+        <SelectContent position="popper" side="left">
+          {DISPLAY_MODE_OPTIONS.map(opt => {
+            const OptIcon = DISPLAY_MODE_ICONS[opt.value]
+            return (
+              <SelectItem key={opt.value} value={opt.value}>
+                <OptIcon className="size-3.5 text-muted-foreground" />
+                {opt.label}
+              </SelectItem>
+            )
+          })}
+        </SelectContent>
+      </Select>
+      <AddPropertyValueInput displayMode={displayMode} value={newValue} onChange={setNewValue} onKeyDown={handleKeyDown} vaultStatuses={vaultStatuses} />
+      <Button
+        size="icon-xs" onClick={() => onAdd(newKey, newValue, displayMode)}
+        disabled={!newKey.trim()} title="Add property"
+        data-testid="add-property-confirm"
+      >
+        <Check className="size-3.5" />
+      </Button>
+      <Button size="icon-xs" variant="outline" onClick={onCancel} title="Cancel" data-testid="add-property-cancel">
+        <X className="size-3.5" />
+      </Button>
+    </div>
+  )
+}
